Progress Made on a 1.7 update... Quitting to make alt less complex plugin that doesn't use NBT...

+++ src/main/java/com/onarandombox/multiverseinventories/api/PlayerStats.java
@@ -20,7 +20,7 @@ public class PlayerStats {
     /**
      * Default health value.
      */
-    public static final int HEALTH = 20;
+    public static final double HEALTH = 20;
     /**
      * Default experience value.
      */
